feat: withdraw without NFT with Goerli issue
