+++ src/pages/WalletPage/index.tsx
@@ -3,18 +3,13 @@ import { Container, Row, Col, Button } from 'react-bootstrap'
 import { getUserBalances } from '../../api/balances-api';
 import { Balance } from '@imtbl/core-sdk'
 import { utils, BigNumber } from "ethers";
-import { Link, ImmutableXClient, ImmutableMethodResults, ERC721TokenType, ETHTokenType, ImmutableRollupStatus } from '@imtbl/imx-sdk';
+import { Link } from '@imtbl/imx-sdk';
 import { useWalletHook } from '../../components/NavBar/useWallethook';
 import Wallet from '../../components/NavBar/Wallet';
 import WalletFeatures from '../../components/WalletFeatures/WalletFeatures';
 
-interface BridgingProps {
-    client: ImmutableXClient,
-    linked: Link,
-    wallet2: string
-  }
 
-export const WalletPage = ({client, linked, wallet2}: BridgingProps) => {
+export const WalletPage = () => {
     const [balances, setBalances] = useState<Balance[]>([]);
     const wallet = useWalletHook();
 
@@ -36,77 +31,6 @@ export const WalletPage = ({client, linked, wallet2}: BridgingProps) => {
         }
     }, [wallet?.walletInfo?.address])
 
-    // Withdraw Section
-
-
-
-    // withdrawals
-    const [preparingWithdrawals, setPreparingWithdrawals] = useState<ImmutableMethodResults.ImmutableGetWithdrawalsResult>(Object);
-    const [readyWithdrawals, setReadyWithdrawals] = useState<ImmutableMethodResults.ImmutableGetWithdrawalsResult>(Object);
-    const [completedWithdrawals, setCompletedWithdrawals] = useState<ImmutableMethodResults.ImmutableGetWithdrawalsResult>(Object);
-    // eth
-    const [depositAmount, setDepositAmount] = useState('');
-    const [prepareAmount, setPrepareAmount] = useState('');
-    // nft
-    // const [depositTokenId, setDepositTokenId] = useState('');
-    // const [depositTokenAddress, setDepositTokenAddress] = useState('');
-    // const [prepareTokenId, setPrepareTokenId] = useState('');
-    // const [prepareTokenAddress, setPrepareTokenAddress] = useState('');
-    // const [completeTokenId, setCompleteTokenId] = useState('');
-    // const [completeTokenAddress, setCompleteTokenAddress] = useState('');
-    
-    useEffect(() => {
-        load()
-    }, [])
-    
-    async function load(): Promise<void> {
-        setPreparingWithdrawals(await client.getWithdrawals({
-        user: wallet2,
-        rollup_status: ImmutableRollupStatus.included
-        })) // included in batch awaiting confirmation
-        setReadyWithdrawals(await client.getWithdrawals({
-        user: wallet2,
-        rollup_status: ImmutableRollupStatus.confirmed,
-        withdrawn_to_wallet: false
-        })) // confirmed on-chain in a batch and ready to be withdrawn
-        setCompletedWithdrawals(await client.getWithdrawals({
-        user: wallet2,
-        withdrawn_to_wallet: true
-        })) // confirmed on-chain in a batch and already withdrawn to L1 wallet
-    };
-
-    // // prepare an NFT withdrawal
-    // async function prepareWithdrawalNFT() {
-    //     await link.prepareWithdrawal({
-    //     type: ERC721TokenType.ERC721,
-    //     tokenId: prepareTokenId,
-    //     tokenAddress: prepareTokenAddress
-    //     })
-    // };
-
-    // prepare an eth withdrawal
-    async function prepareWithdrawalETH() {
-        await linked.prepareWithdrawal({
-        type: ETHTokenType.ETH,
-        amount: prepareAmount,
-        })
-    };
-
-    // // complete an NFT withdrawal
-    // async function completeWithdrawalNFT() {
-    //     await link.completeWithdrawal({
-    //     type: ERC721TokenType.ERC721,
-    //     tokenId: completeTokenId,
-    //     tokenAddress: completeTokenAddress
-    //     })
-    // };
-
-    // complete an eth withdrawal
-    async function completeWithdrawalETH() {
-        await linked.completeWithdrawal({
-        type: ETHTokenType.ETH,
-        })
-    };
 
     return (
         <>
@@ -131,21 +55,6 @@ export const WalletPage = ({client, linked, wallet2}: BridgingProps) => {
                                 <h3 style={{ marginTop: '50px' }}><strong>History</strong></h3>
                                 <Button variant="info" onClick={showUserHistory}>Click here to see transactions</Button>
                                 <WalletFeatures />
-                                <div>
-                                Prepare ETH for withdrawal (submit to be rolled up and confirmed on chain in the next batch):
-                                <br/>
-                                <label>
-                                    Amount (ETH):
-                                    <input type="text" value={prepareAmount} onChange={e => setPrepareAmount(e.target.value)} />
-                                </label>
-                                <button onClick={prepareWithdrawalETH}>Prepare ETH Withdrawal</button>
-                                </div>
-                                <br/><br/>
-                                <div>
-                                Complete ETH withdrawal (withdraws entire eth balance that is ready for withdrawal to L1 wallet):
-                                <br/>
-                                <button onClick={completeWithdrawalETH}>Complete ETH Withdrawal</button>
-                                </div>
                             </> :
                             <>
                                 {<h3>'You are not signed in'</h3>}
